remove unreliable threading.active_count() check
This also counts both pytest and non-pytest threads that haven't yet
had their __del__ scheduled, which the interpreter seems to do based
on the nr. of code (opcodes) and is thus unreliable.

The reproducer for this was starting a Process-only test (no Thread)
and inserting a specific number of arbitrary python commands.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>